refactor(workflows): extract shared watchdog DAL + admin UI into packages/, unify ActionResult
Removes ~1500 lines of provider/middleman cross-app duplication (per CLAUDE.md
"add to packages/ rather than duplicating across apps").

- packages/db/watchdogStore: single source of the watchdog heal-state SQL,
  parameterized by table. createWatchdogHealStateStore (worker) +
  readWatchdogHealState (UI, typed drizzle select). The UI read now uses the
  typed table instead of raw SELECT * — timestamps share the worker's write
  codec (no tz-less local/UTC drift) and DB errors propagate instead of being
  swallowed as "all schedules healthy".
- packages/ui/components/workflows: shared columns / WorkflowsTable /
  SchedulesTab / WorkflowsTabs / WorkflowDetailClient. Server actions stay per
  app and are passed in as props, with a feedback prop reconciling provider's
  toast vs middleman's inline terminate-error UX.
- packages/ui/lib/actionResult: canonical ActionResult; middleman ported to
  withRequireOwner so both apps share one result shape.
- app DAL/UI files become thin wrappers binding their schema table + actions.
