+++ apps/middleman/src/actions/Workflows.test.ts
@@ -1,7 +1,27 @@
 jest.mock('server-only', () => ({}))
 
 const requireAdmin = jest.fn()
-jest.mock('@/lib/utils/actions', () => ({ requireAdmin: () => requireAdmin() }))
+jest.mock('@/lib/utils/actions', () => {
+  const actionResult = require('@igniter/ui/lib/actionResult')
+  return {
+    requireAdmin: () => requireAdmin(),
+    // Faithful stand-in for the real wrapper: enforces requireAdmin and maps to
+    // the shared ActionResult<T> shape (see @/lib/utils/actions).
+    withRequireOwner: async (action: () => Promise<unknown>) => {
+      try {
+        await requireAdmin()
+        return actionResult.success(await action())
+      } catch (err: unknown) {
+        const message = err instanceof Error ? err.message : 'An unexpected error occurred'
+        const code =
+          message === 'Unauthorized' || message === 'Not logged in'
+            ? 'UNAUTHORIZED'
+            : 'INTERNAL_ERROR'
+        return actionResult.error(code, message)
+      }
+    },
+  }
+})
 
 const terminate = jest.fn().mockResolvedValue(undefined)
 const getHandle = jest.fn(() => ({ terminate }))
@@ -56,16 +76,18 @@ describe('middleman Workflows actions', () => {
     requireAdmin.mockResolvedValue(undefined)
   })
 
-  it('ListWorkflows returns ad-hoc success shape with mapped data', async () => {
+  it('ListWorkflows returns ActionResult success with mapped data', async () => {
     const res = await ListWorkflows({}, { pageIndex: 0, pageSize: 10 })
     expect(res.success).toBe(true)
-    expect(res.data?.items[0].workflowId).toBe('wf-a')
+    if (!res.success) throw new Error('expected success')
+    expect(res.data.items[0].workflowId).toBe('wf-a')
   })
 
   it('GetScheduleHealth reports paused schedules (empty health)', async () => {
     const res = await GetScheduleHealth()
     expect(res.success).toBe(true)
-    expect(res.data?.[0].state).toBe('paused')
+    if (!res.success) throw new Error('expected success')
+    expect(res.data[0].state).toBe('paused')
   })
 
   it('TerminateWorkflow calls handle.terminate', async () => {
@@ -79,7 +101,7 @@ describe('middleman Workflows actions', () => {
     requireAdmin.mockRejectedValue(new Error('Unauthorized'))
     const res = await ListWorkflows({}, { pageIndex: 0, pageSize: 10 })
     expect(res.success).toBe(false)
-    expect(res.error).toBe('Unauthorized')
-    expect(res.data).toBeUndefined()
+    if (res.success) throw new Error('expected failure')
+    expect(res.error.message).toBe('Unauthorized')
   })
 })
