+++ apps/provider/src/lib/dal/watchdogHealState.test.ts
@@ -1,27 +1,27 @@
 jest.mock('server-only', () => ({}))
 
-const execute = jest.fn()
+const from = jest.fn()
 jest.mock('@/db', () => ({
-  getDb: () => ({ execute }),
+  getDb: () => ({ select: () => ({ from }) }),
 }))
 
 import { listWatchdogHealState } from './watchdogHealState'
 
 describe('listWatchdogHealState (provider)', () => {
-  beforeEach(() => execute.mockReset())
+  beforeEach(() => from.mockReset())
 
-  it('maps rows (array driver result), defensively across column casings', async () => {
-    execute.mockResolvedValue([
+  it('maps typed drizzle rows, serializing Date columns to ISO strings', async () => {
+    from.mockResolvedValue([
       {
         scheduleId: 'GovernanceSync-scheduled',
         unstucks: 3,
         injectedTriggers: 1,
-        lastHealTriggerAt: '2026-07-01T00:00:00.000Z',
+        lastHealTriggerAt: new Date('2026-07-01T00:00:00.000Z'),
         lastActionCount: 9,
         unhealthy: true,
-        observed_unhealthy: false,
+        observedUnhealthy: false,
         recreations: 2,
-        lastRecreatedAt: '2026-07-01T00:05:00.000Z',
+        lastRecreatedAt: new Date('2026-07-01T00:05:00.000Z'),
       },
     ])
     const rows = await listWatchdogHealState()
@@ -40,22 +40,28 @@ describe('listWatchdogHealState (provider)', () => {
     ])
   })
 
-  it('supports the {rows} driver shape', async () => {
-    execute.mockResolvedValue({ rows: [{ scheduleId: 's', unstucks: 0, unhealthy: false }] })
+  it('maps null timestamp columns to null', async () => {
+    from.mockResolvedValue([
+      {
+        scheduleId: 's',
+        unstucks: 0,
+        injectedTriggers: 0,
+        lastHealTriggerAt: null,
+        lastActionCount: 0,
+        unhealthy: false,
+        observedUnhealthy: false,
+        recreations: 0,
+        lastRecreatedAt: null,
+      },
+    ])
     const rows = await listWatchdogHealState()
-    expect(rows[0].scheduleId).toBe('s')
-    expect(rows[0].unstucks).toBe(0)
-  })
-
-  it('defaults recreations to 0 and lastRecreatedAt to null when absent', async () => {
-    execute.mockResolvedValue([{ scheduleId: 's', unstucks: 0, unhealthy: false }])
-    const rows = await listWatchdogHealState()
-    expect(rows[0].recreations).toBe(0)
+    expect(rows[0].lastHealTriggerAt).toBeNull()
     expect(rows[0].lastRecreatedAt).toBeNull()
+    expect(rows[0].recreations).toBe(0)
   })
 
-  it('degrades to [] when the table is absent (Part A not merged)', async () => {
-    execute.mockRejectedValue(new Error('relation "watchdog_heal_state" does not exist'))
-    await expect(listWatchdogHealState()).resolves.toEqual([])
+  it('propagates errors (a DB outage must not read as "all schedules healthy")', async () => {
+    from.mockRejectedValue(new Error('relation "watchdog_heal_state" does not exist'))
+    await expect(listWatchdogHealState()).rejects.toThrow('watchdog_heal_state')
   })
 })
