+++ packages/ui/package.json
@@ -34,6 +34,7 @@
     "@cosmjs/stargate": "^0.33.1",
     "@hookform/resolvers": "^4.1.3",
     "@igniter/graphql": "workspace:*",
+    "@igniter/temporal": "workspace:*",
     "@igniter/tx-verify": "workspace:*",
     "@mui/system": "^5.14.11",
     "@poktscan/vault-siwp": "github:trustsoothe/vault-siwp#shannon-adr36",
